Split coop TUI outline and detail rendering
Committed-By-Agent: codex
