+++ pkg/coop/tui/detail.go
@@ -1,0 +1,444 @@
+package tui
+
+import (
+	"strings"
+
+	"charm.land/lipgloss/v2"
+
+	"github.com/stripe/stripe-cli/pkg/coop"
+)
+
+var detailSections = []string{"Summary", "Files", "Checks", "Reference"}
+
+func (m Model) renderDetail() string {
+	if m.session == nil {
+		return ""
+	}
+	if m.selected.kind == navigationChapter {
+		return m.renderChapterDetail(m.selected.chapterIndex)
+	}
+	stepIndex, ok := m.selectedStepIndex()
+	if !ok {
+		return ""
+	}
+	node, err := m.session.NodeByNumber(stepIndex + 1)
+	if err != nil {
+		return ""
+	}
+
+	w, innerW := m.detailWidths()
+
+	var md strings.Builder
+	section := detailSections[m.detailTab%len(detailSections)]
+	currentSnippet := m.sdkSnippetStep == stepIndex && m.sdkSnippet != ""
+
+	switch section {
+	case "Summary":
+		m.writeSummaryDetail(&md, node)
+		m.writeStepSDKSnippetDetail(&md, node, currentSnippet)
+	case "Files":
+		m.writeImplementationDetail(&md, node, false)
+	case "Checks":
+		m.writeReviewCommandDetail(&md, node)
+		m.writeAsyncHandlerCheckDetail(&md, node)
+		m.writeVerificationDetail(&md, node)
+	case "Reference":
+		m.writeSDKReferenceDetail(&md, node, currentSnippet)
+		m.writeAsyncHandlerReferenceDetail(&md, node)
+	}
+
+	if node.State == coop.StepSkipped && node.Activity != "" {
+		md.WriteString("*Skipped: " + node.Activity + "*\n\n")
+	}
+
+	content := strings.TrimSpace(md.String())
+	suffix := m.renderDetailSuffix(node, innerW)
+	if content == "" && suffix == "" {
+		return ""
+	}
+
+	var parts []string
+	if header := m.renderDetailHeader(section); header != "" {
+		parts = append(parts, header)
+	}
+	if content != "" {
+		parts = append(parts, clampLines(m.renderMarkdown(content, innerW), innerW))
+	}
+	if suffix != "" {
+		parts = append(parts, suffix)
+	}
+	body := clampLines(strings.Join(parts, "\n"), innerW)
+	box := m.theme.DetailBoxStyle.Width(w).Render(body)
+	return indentBlock(box, detailIndent)
+}
+
+func (m Model) renderChapterDetail(chapterIndex int) string {
+	if m.session == nil || chapterIndex < 0 || chapterIndex >= len(m.session.Chapters) {
+		return ""
+	}
+	w, innerW := m.detailWidths()
+
+	var md strings.Builder
+	section := detailSections[m.detailTab%len(detailSections)]
+	ch := &m.session.Chapters[chapterIndex]
+	switch section {
+	case "Summary":
+		m.writeChapterSummaryDetail(&md, ch, innerW)
+	case "Files":
+		m.writeChapterFilesDetail(&md, ch)
+	case "Checks":
+		m.writeChapterChecksDetail(&md, ch)
+	case "Reference":
+		m.writeChapterReferenceDetail(&md, ch)
+	}
+
+	content := strings.TrimSpace(md.String())
+	suffix := ""
+	if target, ok := m.selectedReviewTarget(); ok && target.kind == "chapter" {
+		suffix = "\n" + m.attentionWrapped("Waiting for you: c confirm all · r request changes", innerW)
+	}
+	if content == "" && suffix == "" {
+		return ""
+	}
+
+	var parts []string
+	if header := m.renderDetailHeader(section); header != "" {
+		parts = append(parts, header)
+	}
+	if content != "" {
+		if section == "Summary" {
+			parts = append(parts, clampLines(content, innerW))
+		} else {
+			parts = append(parts, clampLines(m.renderMarkdown(content, innerW), innerW))
+		}
+	}
+	if suffix != "" {
+		parts = append(parts, suffix)
+	}
+	body := clampLines(strings.Join(parts, "\n"), innerW)
+	box := m.theme.DetailBoxStyle.Width(w).Render(body)
+	return indentBlock(box, detailIndent)
+}
+
+func (m Model) renderDetailHeader(section string) string {
+	if section == "Summary" {
+		return ""
+	}
+	return lipgloss.NewStyle().
+		Foreground(m.theme.HuePurple400).
+		Bold(true).
+		Render(section)
+}
+
+func (m Model) detailWidths() (int, int) {
+	frameW, _ := m.theme.DetailBoxStyle.GetFrameSize()
+	w := m.outlineRuleWidth()
+	if w < 12 {
+		w = 12
+	}
+	innerW := w - frameW
+	if innerW < 8 {
+		innerW = 8
+	}
+	return w, innerW
+}
+
+func indentBlock(s string, spaces int) string {
+	if spaces <= 0 || s == "" {
+		return s
+	}
+	prefix := strings.Repeat(" ", spaces)
+	lines := strings.Split(s, "\n")
+	for i, line := range lines {
+		lines[i] = prefix + line
+	}
+	return strings.Join(lines, "\n")
+}
+
+func (m Model) detailLanguage() string {
+	lang := m.session.Settings["language"]
+	if lang == "" {
+		lang = "javascript"
+	}
+	return lang
+}
+
+func (m Model) writeSummaryDetail(md *strings.Builder, node *coop.SessionNode) {
+	if node.Description != "" {
+		md.WriteString(node.Description + "\n\n")
+	}
+	if node.ReviewPrompt != "" {
+		md.WriteString("**Confirmation steps:** " + node.ReviewPrompt + "\n\n")
+	}
+	if node.Description == "" && node.ReviewPrompt == "" {
+		md.WriteString("*No summary available for this step.*\n\n")
+	}
+}
+
+func (m Model) writeStepSDKSnippetDetail(md *strings.Builder, node *coop.SessionNode, currentSnippet bool) {
+	if node.Type != coop.NodeAPIRequest || node.Request == nil {
+		return
+	}
+	if currentSnippet {
+		md.WriteString("**SDK example**\n")
+		md.WriteString("```" + m.detailLanguage() + "\n")
+		md.WriteString(m.sdkSnippet + "\n")
+		md.WriteString("```\n\n")
+		return
+	}
+	if stepIndex, ok := m.selectedStepIndex(); ok && m.sdkLoading && m.sdkLoadingStep == stepIndex {
+		md.WriteString("*Loading SDK example...*\n\n")
+	}
+}
+
+func (m Model) writeChapterSummaryDetail(md *strings.Builder, ch *coop.SessionChapter, width int) {
+	wrapWidth := width - 2
+	if wrapWidth < 20 {
+		wrapWidth = 20
+	}
+	md.WriteString("Steps\n")
+	for _, node := range ch.Nodes {
+		md.WriteString("  " + chapterStepStatusLabel(node) + " " + node.Title + "\n")
+	}
+	md.WriteString("\n")
+	if changed := chapterChangedFiles(ch); changed != "" {
+		md.WriteString("Changed\n")
+		for _, line := range strings.Split(wordWrap(changed, wrapWidth), "\n") {
+			md.WriteString("  " + line + "\n")
+		}
+		md.WriteString("\n")
+	}
+	if checks := chapterConfirmationSteps(ch); checks != "" {
+		md.WriteString("Confirmation steps\n")
+		for _, line := range strings.Split(wordWrap(checks, wrapWidth), "\n") {
+			md.WriteString("  " + line + "\n")
+		}
+		md.WriteString("\n")
+	}
+	md.WriteString("Agent help\n")
+	for _, line := range strings.Split(wordWrap("The agent should run relevant checks, keep any app or server available, share a local URL when useful, and create or identify test data.", wrapWidth), "\n") {
+		md.WriteString("  " + line + "\n")
+	}
+	md.WriteString("\n")
+}
+
+func (m Model) writeChapterFilesDetail(md *strings.Builder, ch *coop.SessionChapter) {
+	wrote := false
+	for _, node := range ch.Nodes {
+		if node.Implementation == nil || node.Implementation.File == "" {
+			continue
+		}
+		md.WriteString("- `" + implementationFileLabel(node.Implementation) + "` — " + node.Title + "\n")
+		wrote = true
+	}
+	if wrote {
+		md.WriteString("\n")
+		return
+	}
+	md.WriteString("*No files reported for this section yet.*\n\n")
+}
+
+func (m Model) writeChapterChecksDetail(md *strings.Builder, ch *coop.SessionChapter) {
+	wrote := false
+	for _, node := range ch.Nodes {
+		if node.ReviewPrompt != "" {
+			md.WriteString("- " + node.Title + ": " + node.ReviewPrompt + "\n")
+			wrote = true
+		}
+		for _, verification := range node.Verifications {
+			prefix := "✗"
+			if verification.Passed {
+				prefix = "✓"
+			}
+			md.WriteString("- " + prefix + " " + node.Title + ": " + verification.Check + "\n")
+			wrote = true
+		}
+		if command := reviewCommandForNode(&node); command != "" {
+			md.WriteString("- `" + strings.ReplaceAll(command, "`", "'") + "`\n")
+			wrote = true
+		}
+	}
+	if wrote {
+		md.WriteString("\n")
+		return
+	}
+	md.WriteString("*No confirmation steps reported for this section yet.*\n\n")
+}
+
+func (m Model) writeChapterReferenceDetail(md *strings.Builder, ch *coop.SessionChapter) {
+	wrote := false
+	for _, node := range ch.Nodes {
+		if node.Type == coop.NodeAsyncHandler && len(node.Events) > 0 {
+			md.WriteString("- `" + node.Events[0] + "` webhook trigger for " + node.Title + "\n")
+			wrote = true
+		}
+		if node.Type == coop.NodeAPIRequest && node.Request != nil {
+			md.WriteString("- `" + strings.ToUpper(node.Request.Method) + " " + node.Request.Path + "` for " + node.Title + "\n")
+			wrote = true
+		}
+	}
+	if wrote {
+		md.WriteString("\n")
+		return
+	}
+	md.WriteString("*No reference metadata for this section yet.*\n\n")
+}
+
+func chapterStepStatusLabel(node coop.SessionNode) string {
+	switch node.State {
+	case coop.StepDone:
+		return "✓"
+	case coop.StepActive:
+		return "●"
+	case coop.StepReview:
+		return "◆"
+	case coop.StepSkipped:
+		return "–"
+	default:
+		return "○"
+	}
+}
+
+func chapterChangedFiles(ch *coop.SessionChapter) string {
+	var files []string
+	for _, node := range ch.Nodes {
+		if node.Implementation != nil && node.Implementation.File != "" {
+			files = append(files, implementationFileLabel(node.Implementation))
+		}
+	}
+	return strings.Join(files, ", ")
+}
+
+func chapterConfirmationSteps(ch *coop.SessionChapter) string {
+	var agentChecks []string
+	seenAgentCheck := map[string]bool{}
+	for _, node := range ch.Nodes {
+		for _, verification := range node.Verifications {
+			check := strings.TrimSpace(verification.Check)
+			if !verification.Passed || check == "" || seenAgentCheck[check] {
+				continue
+			}
+			seenAgentCheck[check] = true
+			if node.Title != "" {
+				check = node.Title + ": " + check
+			}
+			agentChecks = append(agentChecks, check)
+		}
+	}
+	if len(agentChecks) > 0 {
+		return strings.Join(agentChecks, " ")
+	}
+
+	var checks []string
+	for _, node := range ch.Nodes {
+		if node.ReviewPrompt != "" {
+			checks = append(checks, node.Title+": "+node.ReviewPrompt)
+		}
+	}
+	return strings.Join(checks, " ")
+}
+
+func (m Model) writeAsyncHandlerCheckDetail(md *strings.Builder, node *coop.SessionNode) {
+	if node.Type != coop.NodeAsyncHandler || len(node.Events) == 0 {
+		return
+	}
+	md.WriteString("**How to verify:**\n\n")
+	md.WriteString("1. `stripe listen --forward-to localhost:<port>/webhook`\n")
+	md.WriteString("2. `stripe trigger " + node.Events[0] + "`\n")
+	md.WriteString("3. Confirm your handler processes the event\n\n")
+}
+
+func (m Model) writeAsyncHandlerReferenceDetail(md *strings.Builder, node *coop.SessionNode) {
+	if node.Type != coop.NodeAsyncHandler || len(node.Events) == 0 {
+		return
+	}
+	md.WriteString("**Webhook trigger:**\n\n")
+	md.WriteString("`stripe trigger " + node.Events[0] + "`\n\n")
+}
+
+func (m Model) writeReviewCommandDetail(md *strings.Builder, node *coop.SessionNode) {
+	command := reviewCommandForNode(node)
+	if command == "" {
+		return
+	}
+	md.WriteString("**Review command:**\n\n")
+	md.WriteString("`" + strings.ReplaceAll(command, "`", "'") + "`\n\n")
+}
+
+func (m Model) writeSDKReferenceDetail(md *strings.Builder, node *coop.SessionNode, currentSnippet bool) {
+	if node.Type != coop.NodeAPIRequest {
+		return
+	}
+	if currentSnippet {
+		md.WriteString("**SDK example**\n")
+		md.WriteString("```" + m.detailLanguage() + "\n")
+		md.WriteString(m.sdkSnippet + "\n")
+		md.WriteString("```\n\n")
+		return
+	}
+	if stepIndex, ok := m.selectedStepIndex(); ok && m.sdkLoading && m.sdkLoadingStep == stepIndex {
+		md.WriteString("*Loading reference...*\n\n")
+	}
+}
+
+func (m Model) writeImplementationDetail(md *strings.Builder, node *coop.SessionNode, currentSnippet bool) {
+	if node.Implementation == nil {
+		return
+	}
+	if currentSnippet {
+		md.WriteString("---\n\n")
+	}
+	imp := node.Implementation
+	md.WriteString("**Agent wrote:** `" + implementationFileLabel(imp) + "`\n\n")
+	if imp.Snippet != "" {
+		md.WriteString("```" + m.detailLanguage() + "\n")
+		md.WriteString(imp.Snippet + "\n")
+		md.WriteString("```\n\n")
+	}
+	if imp.Note != "" {
+		md.WriteString("> " + imp.Note + "\n\n")
+	}
+}
+
+func implementationFileLabel(imp *coop.Implementation) string {
+	if imp.File == "" {
+		return ""
+	}
+	if imp.Lines == "" {
+		return imp.File
+	}
+	return imp.File + ":" + imp.Lines
+}
+
+func (m Model) writeVerificationDetail(md *strings.Builder, node *coop.SessionNode) {
+	if len(node.Verifications) == 0 {
+		return
+	}
+	for _, v := range node.Verifications {
+		if v.Passed {
+			md.WriteString("- ✓ " + v.Check + "\n")
+		} else {
+			md.WriteString("- ✗ " + v.Check + "\n")
+		}
+	}
+	md.WriteString("\n")
+}
+
+func (m Model) renderDetailSuffix(node *coop.SessionNode, width int) string {
+	var suffix string
+	if target, ok := m.selectedReviewTarget(); ok && target.kind == "step" && node.State == coop.StepReview {
+		suffix = "\n" + m.attentionWrapped("Waiting for you: c confirm · r request changes", width)
+	}
+	return suffix
+}
+
+func (m Model) attentionWrapped(text string, width int) string {
+	if width < 1 {
+		width = 1
+	}
+	lines := strings.Split(wordWrap(text, width), "\n")
+	for i, line := range lines {
+		lines[i] = m.theme.AttentionStyle.Render(line)
+	}
+	return strings.Join(lines, "\n")
+}
